requirements.txt
Flask==2.3.2
transformers==4.30.2
torch==2.0.1
pandas==2.0.3
numpy==1.24.3
openai==0.27.8
python-dotenv==1.0.0